Pouvoir lister les produits selon les catégories

+++ src/main/java/fr/epsi/mspr/repositories/ProductRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long> {
 
-    List<Product> findByCategoryId(int category_id);
+    List<Product> findByCategoryId(Long category_id);
 }
